UP: move satisfied watched clauses
Satisfied watched clauses of a literal that is propagated could be
processed by either updating the blocking literal for the current
clause, or by moving them to the watch list of the satisfied literal.

The former means less work right now, and only little work in future
iterations.

Latter means more work right now, and no work in future iterations, as
the clause is not present in the list anymore.

The change results in clauses being processed in different order, and
hence it's run time performance can differ, as the search also differs.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>